# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
  details?: string;
}

export interface TikTokMetric {
  label: string;
  value: string | number | boolean | null;
  available: boolean;
}

export type TikTokStoryRaw = { [key: string]: unknown };

export interface TikTokStory {
  id: string;
  caption?: string;
  thumbnailUrl?: string;
  videoUrl?: string;
  createdAt?: string;
  views?: number;
  likes?: number;
  shares?: number;
  comments?: number;
  raw?: TikTokStoryRaw;
}

export type TikTokProfileRaw = { [key: string]: unknown };

export interface TikTokProfile {
  username: string;
  displayName?: string;
  avatarUrl?: string;
  bio?: string;
  verified?: boolean;
  privateAccount?: boolean;
  region?: string;
  language?: string;
  accountLevel?: string;
  followers?: number;
  following?: number;
  likes?: number;
  videos?: number;
  friends?: number;
  createdAt?: string;
  updatedAt?: string;
  contactEmail?: string;
  contactPhone?: string;
  profileUrl?: string;
  raw?: TikTokProfileRaw;
}

export type TikTokProfileResultRawSummary = { [key: string]: unknown };

export interface TikTokProfileResult {
  profile: TikTokProfile;
  stories: TikTokStory[];
  metrics: TikTokMetric[];
  flags: TikTokMetric[];
  source: string;
  generatedAt: string;
  missingFields: string[];
  rawSummary?: TikTokProfileResultRawSummary;
}

export type GetTikTokProfileParams = {
  /**
   * TikTok username without @
   * @minLength 1
   * @maxLength 64
   */
  username: string;
};