# lib/api-zod/src/generated/types/tikTokMetric.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TikTokMetric {
  label: string;
  value: string | number | boolean | null;
  available: boolean;
}