# lib/api-zod/src/generated/types/tikTokProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TikTokProfileRaw } from "./tikTokProfileRaw";

export interface TikTokProfile {
  username: string;
  displayName?: string;
  avatarUrl?: string;
  bio?: string;
  verified?: boolean;
  privateAccount?: boolean;
  region?: string;
  language?: string;
  accountLevel?: string;
  followers?: number;
  following?: number;
  likes?: number;
  videos?: number;
  friends?: number;
  createdAt?: string;
  updatedAt?: string;
  contactEmail?: string;
  contactPhone?: string;
  profileUrl?: string;
  raw?: TikTokProfileRaw;
}