# lib/api-zod/src/generated/types/tikTokProfileResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TikTokMetric } from "./tikTokMetric";
import type { TikTokProfile } from "./tikTokProfile";
import type { TikTokProfileResultRawSummary } from "./tikTokProfileResultRawSummary";
import type { TikTokStory } from "./tikTokStory";

export interface TikTokProfileResult {
  profile: TikTokProfile;
  stories: TikTokStory[];
  metrics: TikTokMetric[];
  flags: TikTokMetric[];
  source: string;
  generatedAt: string;
  missingFields: string[];
  rawSummary?: TikTokProfileResultRawSummary;
}