# lib/api-zod/src/generated/types/tikTokStory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TikTokStoryRaw } from "./tikTokStoryRaw";

export interface TikTokStory {
  id: string;
  caption?: string;
  thumbnailUrl?: string;
  videoUrl?: string;
  createdAt?: string;
  views?: number;
  likes?: number;
  shares?: number;
  comments?: number;
  raw?: TikTokStoryRaw;
}